Refactor DatasetsContent to reorganize hooks into a dedicated datasets directory and remove unused hooks for improved code structure and maintainability
