+++ DashAI/front/src/hooks/datasets/useDatasetFlow.js
@@ -1,6 +1,6 @@
-import { startJobPolling } from "../utils/jobPoller";
-import { enqueueDatasetJob } from "../api/job";
-import { createDataset, deleteDataset } from "../api/datasets";
+import { startJobPolling } from "../../utils/jobPoller";
+import { enqueueDatasetJob } from "../../api/job";
+import { createDataset, deleteDataset } from "../../api/datasets";
 
 export function useDatasetFlow({
   datasets,
