+++ DashAI/front/src/hooks/datasets/useDatasets.js
@@ -5,8 +5,8 @@ import {
   getDatasetInfo,
   updateDataset,
   createDataset,
-} from "../api/datasets";
-import { startJobPolling } from "../utils/jobPoller";
+} from "../../api/datasets";
+import { startJobPolling } from "../../utils/jobPoller";
 import { replace } from "formik";
 
 export function useDatasets({ enqueueSnackbar, t }) {
